Fixed demo and box crates

+++ demo/src/main.rs
@@ -11,10 +11,7 @@ fn main() {
             DefaultPlugins
                 .set(WindowPlugin {
                     primary_window: Some(Window {
-                        resolution: WindowResolution::new(
-                            (768.0 * 1.5) as u32,
-                            (560.0 * 1.5) as u32,
-                        ),
+                        resolution: WindowResolution::new(768.0 * 1.5, 560.0 * 1.5),
                         ..Default::default()
                     }),
                     ..Default::default()
@@ -22,8 +19,6 @@ fn main() {
                 .set(ImagePlugin::default_nearest()),
             bevy_pretty_box::PrettyBoxPlugin,
             bevy_seedling::SeedlingPlugin::default(),
-            // bevy_inspector_egui::bevy_egui::EguiPlugin::default(),
-            // bevy_inspector_egui::quick::WorldInspectorPlugin::default(),
         ))
         .add_plugins((
             audio::AudioPlugin,
